Remove instructions to add post-deploy hook
As I was following along with this tutorial, I used the instructions for the post-deploy hook to have it automatically run the generate command. But I noticed this warning in my terminal.

```
Warning: The `prisma generate` command was executed twice. Since Prisma 1.31, the Prisma client is generated automatically after running `prisma deploy`. It is not necessary to generate it via a `post-deploy` hook any more, you can therefore remove the hook if you do not need it otherwise.
```

I think removing this would keep the tutorial more up to date with the current state of the tooling.